+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 ## [Unreleased]
 - Asynchronous email sending
+
+## [0.0.2] - 30-05-2019
+### Added
+- Admin site header
 - Admin CRM module
+- Add field contacts to campaign
+
+### Changed
+- Alter field campaign on campaigncontact
 
 ## [0.0.1] - 30-05-2019
 ### Added